+++ node/mdSchema.ts
@@ -22,6 +22,7 @@ export const ORGANIZATION_FIELDS = [
   'costCenters',
   'status',
   'created',
+  'customFields',
 ]
 export const ORGANIZATION_SCHEMA_VERSION = 'v0.0.7'
 
@@ -36,7 +37,15 @@ export const COST_CENTER_FIELDS = [
   'businessDocument',
   'customFields',
 ]
-export const COST_CENTER_SCHEMA_VERSION = 'v0.0.6'
+export const COST_CENTER_SCHEMA_VERSION = 'v0.0.7'
+
+export const B2B_SETTINGS_DATA_ENTITY = 'b2b_settings'
+export const B2B_SETTINGS_FIELDS = [
+  'autoApprove',
+  'defaultPaymentTerms',
+  'defaultPriceTables',
+]
+export const B2B_SETTINGS_SCHEMA_VERSION = 'v0.0.8'
 
 export const schemas = [
   {
@@ -124,6 +133,10 @@ export const schemas = [
           title: 'Created',
           format: 'date-time',
         },
+        customFields: {
+          type: ['array', 'null'],
+          title: 'Custom Fields',
+        },
       },
       'v-indexed': ['name', 'status', 'created'],
       'v-immediate-indexing': true,
@@ -159,12 +172,34 @@ export const schemas = [
           type: ['string', 'null'],
           title: 'Phone Number',
         },
-        // customFields: {
-        //   type: ['array', 'null'],
-        //   title: 'Custom Fields',
-        // },
+        customFields: {
+          type: ['array', 'null'],
+          title: 'Custom Fields',
+        },
       },
       'v-indexed': ['name', 'organization', 'businessDocument'],
+      'v-immediate-indexing': true,
+      'v-cache': false,
+    },
+  },
+  {
+    name: B2B_SETTINGS_DATA_ENTITY,
+    version: B2B_SETTINGS_SCHEMA_VERSION,
+    body: {
+      properties: {
+        autoApprove: {
+          type: 'boolean',
+          title: 'Auto Approve',
+        },
+        defaultPaymentTerms: {
+          type: 'array',
+          title: 'Default Payment Terms',
+        },
+        defaultPriceTables: {
+          type: 'array',
+          title: 'Default Price Tables',
+        },
+      },
       'v-immediate-indexing': true,
       'v-cache': false,
     },