+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Allow users to set auto approval for new organization with a new setting fields
+
 ## [0.22.0] - 2022-09-16
 
 ### Added
@@ -36,6 +40,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Added the cookie header to storefront permissions requests passing the authentication token.
+
+### Added
+
+- Allow users to set custom fields to hold information about organizations
+
+## [0.20.0] - 2022-08-30
+
+## [Unreleased]
+
+### Added
+
+- Allow users to set auto approval for new organization with a new setting fields
 
 ## [0.19.7] - 2022-08-23